Fix refactoring of transformers
