+++ modules/AppearanceAPI/src/main/java/org/gephi/appearance/AttributeRankingImpl.java
@@ -83,7 +83,10 @@ public class AttributeRankingImpl extends RankingImpl {
     @Override
     public boolean isValid(Graph graph) {
         Column col = column.get();
-        return col != null && col.getIndex() != -1;
+        if(col != null && col.getIndex() != -1) {
+            return col.isNumber();
+        }
+        return false;
     }
 
     @Override
