+++ modules/AppearancePluginUI/src/main/java/org/gephi/ui/appearance/plugin/RankingLabelSizeTransformerUI.java
@@ -40,15 +40,15 @@
  Portions Copyrighted 2013 Gephi Consortium.
  */
 
-package org.gephi.ui.appearance.plugin.unique;
+package org.gephi.ui.appearance.plugin;
 
 import javax.swing.AbstractButton;
 import javax.swing.Icon;
 import javax.swing.JPanel;
 import org.gephi.appearance.api.Function;
-import org.gephi.appearance.api.SimpleFunction;
-import org.gephi.appearance.plugin.UniqueLabelSizeTransformer;
-import org.gephi.appearance.spi.SimpleTransformer;
+import org.gephi.appearance.api.RankingFunction;
+import org.gephi.appearance.plugin.RankingLabelSizeTransformer;
+import org.gephi.appearance.spi.RankingTransformer;
 import org.gephi.appearance.spi.TransformerCategory;
 import org.gephi.appearance.spi.TransformerUI;
 import org.gephi.ui.appearance.plugin.category.DefaultCategory;
@@ -58,24 +58,14 @@ import org.openide.util.lookup.ServiceProvider;
 /**
  * @author mbastian
  */
-@ServiceProvider(service = TransformerUI.class, position = 700)
-public class UniqueLabelSizeTransformerUI implements TransformerUI {
+@ServiceProvider(service = TransformerUI.class, position = 800)
+public class RankingLabelSizeTransformerUI implements TransformerUI {
 
-    private UniqueSizeTransformerPanel panel;
-
-    @Override
-    public String getDisplayName() {
-        return NbBundle.getMessage(UniqueLabelSizeTransformerUI.class, "Unique.name");
-    }
+    private RankingSizeTransformerPanel panel;
 
     @Override
     public TransformerCategory getCategory() {
         return DefaultCategory.LABEL_SIZE;
-    }
-
-    @Override
-    public String getDescription() {
-        return null;
     }
 
     @Override
@@ -84,11 +74,21 @@ public class UniqueLabelSizeTransformerUI implements TransformerUI {
     }
 
     @Override
+    public String getDisplayName() {
+        return NbBundle.getMessage(RankingLabelSizeTransformerUI.class, "Attribute.ranking.name");
+    }
+
+    @Override
+    public String getDescription() {
+        return null;
+    }
+
+    @Override
     public synchronized JPanel getPanel(Function function) {
         if (panel == null) {
-            panel = new UniqueSizeTransformerPanel();
+            panel = new RankingSizeTransformerPanel();
         }
-        panel.setup((SimpleFunction) function);
+        panel.setup((RankingFunction) function);
         return panel;
     }
 
@@ -98,7 +98,7 @@ public class UniqueLabelSizeTransformerUI implements TransformerUI {
     }
 
     @Override
-    public Class<? extends SimpleTransformer> getTransformerClass() {
-        return UniqueLabelSizeTransformer.class;
+    public Class<? extends RankingTransformer> getTransformerClass() {
+        return RankingLabelSizeTransformer.class;
     }
 }
