--- modules/AppearancePluginUI/src/main/resources/org/gephi/ui/partition/plugin/Bundle.properties
@@ -1,7 +1,0 @@
-NodeColorTransformerBuilder.ui.name = Color
-NodeColorTransformerPanel.action.randomize = Randomize colors
-NodeColorTransformerPanel.action.allBlacks = All Blacks
-
-EdgeColorTransformerBuilder.ui.name = Color
-EdgeColorTransformerPanel.action.randomize = Randomize
-EdgeColorTransformerPanel.action.allBlacks = All Blacks
