--- modules/AppearancePluginUI/src/main/resources/org/gephi/ui/partition/plugin/Bundle_ca.properties
@@ -1,7 +1,0 @@
-NodeColorTransformerBuilder.ui.name = Color
-NodeColorTransformerPanel.action.randomize = Tria els colors aleatòriament
-NodeColorTransformerPanel.action.allBlacks = Tots negres
-
-EdgeColorTransformerBuilder.ui.name = Color
-EdgeColorTransformerPanel.action.randomize = Fes aleatori
-EdgeColorTransformerPanel.action.allBlacks = Tots negres
