--- modules/AppearancePluginUI/src/main/resources/org/gephi/ui/partition/plugin/Bundle_cs.properties
@@ -1,7 +1,0 @@
-NodeColorTransformerBuilder.ui.name = Barva
-NodeColorTransformerPanel.action.randomize = Barvy náhodn\u011b
-NodeColorTransformerPanel.action.allBlacks = V\u0161echno \u010derné
-
-EdgeColorTransformerBuilder.ui.name = Barva
-EdgeColorTransformerPanel.action.randomize = Náhodn\u011b
-EdgeColorTransformerPanel.action.allBlacks = V\u0161echno \u010derné
