--- modules/AppearancePluginUI/src/main/resources/org/gephi/ui/partition/plugin/Bundle_de.properties
@@ -1,7 +1,0 @@
-NodeColorTransformerBuilder.ui.name = Farbe
-NodeColorTransformerPanel.action.randomize = Wähle Farben zufällig
-NodeColorTransformerPanel.action.allBlacks = Alles schwarz
-
-EdgeColorTransformerBuilder.ui.name = Farbe
-EdgeColorTransformerPanel.action.randomize = Randomisieren
-EdgeColorTransformerPanel.action.allBlacks = Alles schwarz
