--- modules/AppearancePluginUI/src/main/resources/org/gephi/ui/partition/plugin/Bundle_es.properties
@@ -1,7 +1,0 @@
-NodeColorTransformerBuilder.ui.name = Color
-NodeColorTransformerPanel.action.randomize = Colores aleatorios
-NodeColorTransformerPanel.action.allBlacks = Todos negros
-
-EdgeColorTransformerBuilder.ui.name = Color
-EdgeColorTransformerPanel.action.randomize = Colores aleatorios
-EdgeColorTransformerPanel.action.allBlacks = Todos negros
