--- modules/AppearancePluginUI/src/main/resources/org/gephi/ui/partition/plugin/Bundle_fr.properties
@@ -1,7 +1,0 @@
-NodeColorTransformerBuilder.ui.name = Couleur
-NodeColorTransformerPanel.action.randomize = Générer les couleurs aléatoirement
-NodeColorTransformerPanel.action.allBlacks = Tout en noir
-
-EdgeColorTransformerBuilder.ui.name = Couleur
-EdgeColorTransformerPanel.action.randomize = Aléatoire
-EdgeColorTransformerPanel.action.allBlacks = Tout en noir
