--- modules/AppearancePluginUI/src/main/resources/org/gephi/ui/partition/plugin/Bundle_he.properties
@@ -1,6 +1,0 @@
-NodeColorTransformerBuilder.ui.name=\u05e6\u05d1\u05e2
-NodeColorTransformerPanel.action.randomize=Randomize colors
-NodeColorTransformerPanel.action.allBlacks=All Blacks
-EdgeColorTransformerBuilder.ui.name=\u05e6\u05d1\u05e2
-EdgeColorTransformerPanel.action.randomize=Randomize
-EdgeColorTransformerPanel.action.allBlacks=All Blacks
