--- modules/AppearancePluginUI/src/main/resources/org/gephi/ui/partition/plugin/Bundle_ja.properties
@@ -1,7 +1,0 @@
-NodeColorTransformerBuilder.ui.name = \u8272
-NodeColorTransformerPanel.action.randomize = \u8272\u3092\u30e9\u30f3\u30c0\u30e0\u5316
-NodeColorTransformerPanel.action.allBlacks = \u3059\u3079\u3066\u9ed2
-
-EdgeColorTransformerBuilder.ui.name = \u8272
-EdgeColorTransformerPanel.action.randomize = \u30e9\u30f3\u30c0\u30e0\u5316
-EdgeColorTransformerPanel.action.allBlacks = \u3059\u3079\u3066\u9ed2
