--- modules/AppearancePluginUI/src/main/resources/org/gephi/ui/partition/plugin/Bundle_nl.properties
@@ -1,6 +1,0 @@
-NodeColorTransformerBuilder.ui.name=Kleur
-NodeColorTransformerPanel.action.randomize=Willekeurige kleuren
-NodeColorTransformerPanel.action.allBlacks=All Blacks
-EdgeColorTransformerBuilder.ui.name=Kleur
-EdgeColorTransformerPanel.action.randomize=Willekeurig
-EdgeColorTransformerPanel.action.allBlacks=All Blacks
