--- modules/AppearancePluginUI/src/main/resources/org/gephi/ui/partition/plugin/Bundle_pt_BR.properties
@@ -1,7 +1,0 @@
-NodeColorTransformerBuilder.ui.name = Cor
-NodeColorTransformerPanel.action.randomize = Cores aleatórias
-NodeColorTransformerPanel.action.allBlacks = Todos negros
-
-EdgeColorTransformerBuilder.ui.name = Cor
-EdgeColorTransformerPanel.action.randomize = Cores aleatórias
-EdgeColorTransformerPanel.action.allBlacks = Todos negros
