--- modules/AppearancePluginUI/src/main/resources/org/gephi/ui/partition/plugin/Bundle_ru.properties
@@ -1,7 +1,0 @@
-NodeColorTransformerBuilder.ui.name = \u0426\u0432\u0435\u0442
-NodeColorTransformerPanel.action.randomize = \u0421\u043b\u0443\u0447\u0430\u0439\u043d\u044b\u0435 \u0446\u0432\u0435\u0442\u0430
-NodeColorTransformerPanel.action.allBlacks = \u0412\u0441\u0435 \u0447\u0451\u0440\u043d\u044b\u0435
-
-EdgeColorTransformerBuilder.ui.name = \u0426\u0432\u0435\u0442
-EdgeColorTransformerPanel.action.randomize = \u0421\u043b\u0443\u0447\u0430\u0439\u043d\u044b\u0435
-EdgeColorTransformerPanel.action.allBlacks = \u0412\u0441\u0435 \u0447\u0451\u0440\u043d\u044b\u0435
