--- modules/AppearancePluginUI/src/main/resources/org/gephi/ui/partition/plugin/Bundle_tr.properties
@@ -1,6 +1,0 @@
-NodeColorTransformerBuilder.ui.name=Renk
-NodeColorTransformerPanel.action.randomize=Randomize colors
-NodeColorTransformerPanel.action.allBlacks=All Blacks
-EdgeColorTransformerBuilder.ui.name=Renk
-EdgeColorTransformerPanel.action.randomize=Randomize
-EdgeColorTransformerPanel.action.allBlacks=All Blacks
