--- modules/AppearancePluginUI/src/main/resources/org/gephi/ui/partition/plugin/Bundle_zh_CN.properties
@@ -1,7 +1,0 @@
-NodeColorTransformerBuilder.ui.name = \u989c\u8272
-NodeColorTransformerPanel.action.randomize = \u968f\u673a\u989c\u8272
-NodeColorTransformerPanel.action.allBlacks = \u5168\u9ed1
-
-EdgeColorTransformerBuilder.ui.name = \u989c\u8272
-EdgeColorTransformerPanel.action.randomize = \u968f\u673a
-EdgeColorTransformerPanel.action.allBlacks = \u5168\u9ed1
