--- modules/AppearancePluginUI/src/main/resources/org/gephi/ui/partition/plugin/Bundle_zh_TW.properties
@@ -1,6 +1,0 @@
-NodeColorTransformerBuilder.ui.name=Color
-NodeColorTransformerPanel.action.randomize=Randomize colors
-NodeColorTransformerPanel.action.allBlacks=All Blacks
-EdgeColorTransformerBuilder.ui.name=Color
-EdgeColorTransformerPanel.action.randomize=Randomize
-EdgeColorTransformerPanel.action.allBlacks=All Blacks
